execAppend for grouping with top of stack
This will aid in expressing command execution which is more stream-like;
the programmer doesn't know the entire command before running some of
it. Imagine user input: the user pulls on something with the mouse, and
at each movement we run a command, but we want the undo frame to include
everything that was done during that pull.